don't enforce a specific WindowsTargetPlatformVersion
Omitting this setting causes Visual Studio to default to its preferred
one. For Visual Studio 2017, it's 8.1, and 2019/2022 default to the
latest installed 10.0 version.

The `_WIN32_WINNT` hack then causes issues with more recent versions of
the Windows 11 SDK, while not seeming to have any effect on the MinGW
build. After replacing it with an explicit `#define` of
`ENABLE_VIRTUAL_TERMINAL_PROCESSING`, which is not defined in the 8.1
SDK, the projects now compile on default installations of Visual Studio
2017, 2019, and 2022.