disable vt mode if stdout is redirected
`read_cursor_pos()` sends `"\x1b[6n"` to stdout and then expects to
read the cursor position from stdin. This can never work if stdout is
not a console, and previously caused MS-DOS Player to hang in an
infinite loop in that case.

Since `use_vt` is set based on the Windows version and the `-vt` flag
merely toggles this default, we also got different behavior depending
on the OS: On Windows ≥10, `msdos prog >NUL` would hang and
`msdos prog -vt >NUL` would run, but it's the other way around on
earlier Windows versions.